Fixed player ability pool to reduce current value during casting

+++ Assets/Scripts/Player/Behaviours/IAbility.cs
@@ -1,7 +1,7 @@
 ﻿public interface IAbility
 {
     float TickTime { get; }
-    float Cost { get; }
+    int Cost { get; }
     void StartedCasting();
     void Tick();
     void FinishedCasting();
